+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Small internal changes to accomodate a few more clippy lints
 - Update `ratatui` and `ratatui-image` git dependencies to latest upstream
 - Move `ratatui-image/vb64` support under `nightly` feature, enabled by default
+- Fixed a bug where jumping to a page out of range could result in weird `esc` key behavior
+- Added CI
 
 # v0.1.0